docs/mcp-server/requirements.txt
# MCP Server Dependencies
mcp>=0.9.0
aiohttp>=3.9.0
openai>=1.0.0
pillow>=10.0.0
python-dotenv>=1.0.0
